requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
Cerberus==1.3.4
cfgv==3.3.1
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
distlib==0.3.7
Django==4.2.3
django-celery==3.1.17
django-celery-beat==2.5.0
django-timezone-field==5.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
filelock==3.12.2
fpdf==1.7.2
gunicorn==21.2.0
identify==2.5.26
kombu==5.3.1
lxml==4.9.3
mysqlclient==2.2.0
nodeenv==1.8.0
numpy==1.24.4
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
platformdirs==3.9.1
pre-commit==3.3.3
prompt-toolkit==3.0.39
pycryptodome==3.18.0
PyMuPDF==1.22.5
pypdf==3.12.2
PyPDF2==3.0.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
vine==5.0.0
virtualenv==20.24.1
wcwidth==0.2.6
